pyproject.toml
[tool.poetry]
name = "gaming-highlight-gen"
version = "0.1.0"
description = "AI-based gaming highlight generator"
authors = ["Gaming Highlight Gen <dev@example.com>"]
readme = "README.md"
packages = [{ include = "gaming_highlight_gen" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7"
pydantic-settings = "^2.3"
typer = { extras = ["all"], version = "^0.12" }
rich = "^13.7"
structlog = "^24.1"
pyyaml = "^6.0"
ffmpeg-python = "^0.2"
librosa = "^0.10"
soundfile = "^0.12"
opencv-python-headless = "^4.10"
numpy = "^1.26"
scipy = "^1.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-cov = "^5.0"
pytest-mock = "^3.14"
ruff = "^0.4"
mypy = "^1.10"
pre-commit = "^3.7"

[tool.poetry.scripts]
highlight-gen = "gaming_highlight_gen.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=gaming_highlight_gen --cov-report=term-missing --cov-fail-under=80"
